+++ changelog.md
@@ -1,6 +1,7 @@
 # 2.1.0
 - 2.2081 support
 - Cleanup code
+- Fix border flicker
 
 # 2.0.23
 - Fix a small oversight that could break transitions